Feat: (bw) Match BW output with Demul input
We use Armadillo wrapper to help handle the memory layout. In BW, we use an
Armadillo cube to point to the ul_beam_mem. Thus, the data are naturally stored
during computation step. Then, in the beginning of the Demul (equalization part),
we use the exact format to read out memories.

Result: reduce BW time from 0.039 ms to 0.012 ms. (data distribution time = ~0)