Feat: (bw/equal) Remove Arma for data passing
For the primary, avx-512 based solution, we do not use Armadillo but use raw
pointer whenever possible to accelerate. Since C is row-major and Armadillo is
column-major, the modified data layout will differ so we have to modify the
interface of both procuder and consumer doer. The Armadillo based solution is
kept as the fall-back option.